bsze3_sql/requirements.txt
mypy==1.9.0
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
typing_extensions==4.10.0
